% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{check_geo_service}
\alias{check_geo_service}
\title{Runs a basic check to see if
a valid response is returned by
Geoserver and returns TRUE/FALSE}
\usage{
check_geo_service()
}
\description{
Used in unit tests to determine if
tests should be run
}
