% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{check_service}
\alias{check_service}
\title{Runs a basic check to see if
a valid response is returned by
the NPN Portal service
and returns TRUE/FALSE}
\usage{
check_service()
}
\description{
Used in unit tests to determine if
tests should be run
}
