% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npn_phenophases.R
\name{npn_abundance_categories}
\alias{npn_abundance_categories}
\title{Get Abundance Categories}
\usage{
npn_abundance_categories(...)
}
\arguments{
\item{...}{Currently unused.}
}
\value{
A data frame listing all abundance/intensity categories and their
corresponding values.
}
\description{
Gets data on all abundance/intensity categories and includes a data frame of
applicable abundance/intensity values for each category
}
\examples{
\dontrun {
ac <- npn_abundance_categories()
}
}
