% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npn_geoserver.R
\name{npn_check_point_cached}
\alias{npn_check_point_cached}
\title{Check Point Cached}
\usage{
npn_check_point_cached(layer, lat, long, date)
}
\arguments{
\item{layer}{The name of the queried layer.}

\item{lat}{The latitude of the queried point.}

\item{long}{The longitude of the queried point.}

\item{date}{The queried date.}
}
\value{
The numeric value of the cell located at the specified coordinates and date if the value has been queried, otherwise NULL.
}
\description{
Checks in the global variable "point values" to see if the exact data point being requested
has already been asked for and returns the value if it's already saved.
}
\keyword{internal}
