% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npn_dataset.R
\name{npn_datasets}
\alias{npn_datasets}
\title{Get Datasets}
\usage{
npn_datasets(...)
}
\arguments{
\item{...}{Currently unused.}
}
\value{
tibble of datasets and their IDs.
}
\description{
Returns a complete list of information about all datasets integrated into the
NPN dataset. Data can then be pulled for individual datasets using their
unique IDs.
}
\examples{
\dontrun{
npn_datasets()
}
}
