% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npn_data_download.R
\name{npn_get_download_url}
\alias{npn_get_download_url}
\title{Generate Download URL}
\usage{
npn_get_download_url(endpoint)
}
\arguments{
\item{endpoint}{The service point, e.g. "observations/getObservations.json?"}

\item{query_vars}{List of query params}
}
\value{
The URL, as a string
}
\description{
Utility function to create the service point URL. Base URL comes from zzz.R, endpoint is specified in the code, and query_vars should be a list of parameters.
This function will manually put those query parameters into the proper GET syntax.
}
\keyword{internal}
