% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npn_geoserver.R
\name{npn_get_point_data}
\alias{npn_get_point_data}
\title{Get Point Data Value}
\usage{
npn_get_point_data(layer, lat, long, date, store_data = TRUE)
}
\description{
This function can get point data about any of the NPN geospatial layers.
}
\details{
Please note that this function pulls this from the NPN's WCS service so the data may not be totally precise. If
you need precise AGDD values try using the npn_get_agdd_point_data function.
}
