% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npn_partner_groups.R
\name{npn_groups}
\alias{npn_groups}
\title{Get Partner Groups}
\usage{
npn_groups(use_hierarchy = FALSE, ...)
}
\arguments{
\item{use_hierarchy}{Boolean indicating whether or not the list of networks
should be represented in a hierarchy. Defaults to \code{FALSE}.}

\item{...}{Currently unused.}
}
\value{
A tibble of partner groups, including \code{network_id} and
\code{network_name}.
}
\description{
Returns a list of all groups participating in the NPN's data collection
program. These details can be used to further filter other service endpoints'
results.
}
\examples{
\dontrun{
npn_groups()
npn_groups(use_heirarchy = TRUE)
}
}
