% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npn_obsspindloc.R
\name{npn_obsspindloc}
\alias{npn_obsspindloc}
\title{Return all of the data, positive or negative, for an individual or a group
   of individuals, which are of a common species, at any number of locations.}
\usage{
npn_obsspindloc(stationid, speciesid, year = 2010, individualid = NULL, ...)
}
\arguments{
\item{stationid}{Required. Station ID; Use e.g., c(4881, 4882, etc.) if more than
one station desired (numeric).}

\item{speciesid}{Required. Species ID number (numeric).}

\item{year}{Required. Year (numeric). Default: 2010}

\item{individualid}{An individual's ID. (numeric)}

\item{...}{Optional additional curl options (debugging tools mostly)}
}
\value{
Data frame
}
\description{
Return all of the data, positive or negative, for an individual or a group
   of individuals, which are of a common species, at any number of locations.
}
\examples{
\dontrun{
npn_obsspindloc(stationid = c(4881, 4882), speciesid = 3, year = 2009)
npn_obsspindloc(stationid = c(4881, 4882), speciesid = 3)
npn_obsspindloc(stationid = 4881, speciesid = 67)
}
}

