% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npn_phenophases.R
\name{npn_phenophase_definitions}
\alias{npn_phenophase_definitions}
\title{Get Phenophase Definitions}
\usage{
npn_phenophase_definitions(...)
}
\arguments{
\item{...}{Curl options passed on to \code{\link[httr]{GET}}}
}
\description{
Retrieves a complete list of all phenophase definitions.
}
