% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npn_phenophases.R
\name{npn_phenophase_details}
\alias{npn_phenophase_details}
\title{Get Phenophase Details}
\usage{
npn_phenophase_details(ids = NULL, ...)
}
\arguments{
\item{ids}{Takes a vector of phenophase ids for which to retrieve additional
details.}

\item{...}{Currently unused.}
}
\value{
A tibble listing phenophases in the NPN database, including detailed
information for each, filtered by the phenophase ID.
}
\description{
Retrieves additional details for select phenophases, including full list of
applicable phenophase definition IDs and phenophase revision notes over time
}
