% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npn_phenophases.R
\name{npn_phenophases}
\alias{npn_phenophases}
\title{Get Phenophases}
\usage{
npn_phenophases(...)
}
\arguments{
\item{...}{Curl options passed on to \code{\link[httr]{GET}}}
}
\value{
A data frame listing all phenophases available in the NPN database.
}
\description{
Retrieves a complete list of all phenophases in the NPN database.
}
