% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{npn_set_env}
\alias{npn_set_env}
\title{Set Environment}
\usage{
npn_set_env(env = "ops")
}
\arguments{
\item{env}{The environment to use. Should be "ops" or "dev"}
}
\description{
By default this library will call the NPN's production services
but in some cases it's preferable to access the development web services
so this function allows for manually setting the web service endpoints
to use DEV instead. Just pass in "dev" to this function to change the
endpoints to use.
}
