% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npn_get_species.R
\name{npn_species_types}
\alias{npn_species_types}
\title{Get Species Types}
\usage{
npn_species_types(kingdom = c("Animalia", "Plantae"), ...)
}
\arguments{
\item{kingdom}{Character vector of taxonomic kingdoms to filter results by.
Valid values include \code{'Animalia'}, \code{'Plantae'}, or \code{c("Animalia", "Plantae")} (default).}

\item{...}{Currently unused.}
}
\value{
A data frame with a list of the functional types used in the NPN
database, filtered by the specified kingdom.
}
\description{
Return all plant or animal functional types used in the NPN database.
}
\examples{
\dontrun{
npn_species_types("Plantae")
}
}
