% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npn_get_stations.R
\name{npn_stations}
\alias{npn_stations}
\title{Get Station Data}
\usage{
npn_stations(state_code = NULL, ...)
}
\arguments{
\item{state_code}{The postal code of the US state by which to filter
the results returned. Leave empty to get all stations.}

\item{...}{Curl options passed on to \code{\link[httr]{GET}}}
}
\value{
Stations' latitude and longitude, names, and ids.
}
\description{
Get a list of all stations, optionally filtered by state
}
\examples{
\dontrun{
npn_stations()
npn_stations('AZ')
}
}
