% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npn_stations.R
\name{npn_stations_by_location}
\alias{npn_stations_by_location}
\title{Get station data based on a WKT defined geography.}
\usage{
npn_stations_by_location(wkt, ...)
}
\arguments{
\item{wkt}{Required field specifying the WKT geography to use.}

\item{...}{currently unused}
}
\value{
A data frame listing stations filtered based on the WKT geography.
}
\description{
Takes a Well-Known Text based geography as input and returns data for all
stations, including unique IDs, within that boundary.
}
\examples{
\dontrun{
head( npn_stations_by_state(wkt="POLYGON((
-110.94484396954107 32.23623109416672,-110.96166678448247 32.23594069208043,
-110.95960684795904 32.21328646993733,-110.94244071026372 32.21343170728929,
-110.93935080547857 32.23216538049456,-110.94484396954107 32.23623109416672))")
)
}
}
