% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npn_stations.R
\name{npn_stations_by_state}
\alias{npn_stations_by_state}
\title{Get number of stations by state.}
\usage{
npn_stations_by_state(...)
}
\arguments{
\item{...}{Curl options passed on to \code{\link[httr]{GET}}}
}
\value{
A data frame listing stations by state.
}
\description{
Get number of stations by state.
}
\examples{
\dontrun{
head( npn_stations_by_state() )
}
}
