% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npn_stations.R
\name{npn_stations_with_spp}
\alias{npn_stations_with_spp}
\title{Get Stations with Species}
\usage{
npn_stations_with_spp(speciesid, ...)
}
\arguments{
\item{speciesid}{Required. Species id numbers, from 1 to infinity, potentially,
use e.g., `c(52, 53, etc.)` if more than one species desired (numeric).}

\item{...}{Currently unused.}
}
\value{
A data frame with stations' latitude and longitude, names, and ids.
}
\description{
Get a list of all stations which have an individual whom is a member of a
   set of species.
}
\examples{
\dontrun{
npn_stations_with_spp(speciesid = c(52,53,54))
npn_stations_with_spp(speciesid = 53)
}
}
