% Generated by roxygen2 (4.0.2): do not edit by hand
\name{npn_stationsbystate}
\alias{npn_stationsbystate}
\title{Get number of stations by state.}
\usage{
npn_stationsbystate(...)
}
\arguments{
\item{...}{Optional additional curl options (debugging tools mostly)}
}
\value{
Number of stations by state as a data.frame.
}
\description{
Get number of stations by state.
}
\examples{
\donttest{
head( npn_stationsbystate() )
}
}

