% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defunct.R
\name{npn_stationswithspp}
\alias{npn_stationswithspp}
\title{This function renamed to be consistent with other package function names.}
\usage{
npn_stationswithspp(...)
}
\description{
This function renamed to be consistent with other package function names.
}
\keyword{internal}
