% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npn_stationswithspp.R
\name{npn_stationswithspp}
\alias{npn_stationswithspp}
\title{Get a list of all stations which have an individual whom is a member of a
   set of species.}
\usage{
npn_stationswithspp(speciesid, ...)
}
\arguments{
\item{speciesid}{Required. Species id numbers, from 1 to infinity, potentially,
use e.g., c(52, 53, etc.) if more than one species desired (numeric)}

\item{...}{Curl options passed on to \code{\link[httr]{GET}}}
}
\value{
Stations' latitude and longitude, names, and ids.
}
\description{
Get a list of all stations which have an individual whom is a member of a
   set of species.
}
\examples{
\dontrun{
npn_stationswithspp(speciesid = c(52,53,54))
npn_stationswithspp(speciesid = 53)
}
}

