\name{npn_todf}
\alias{npn_todf}
\title{Coerce elements of output from a call to occ to a single data.frame}
\usage{
  npn_todf(x, minimal = FALSE)
}
\arguments{
  \item{x}{An object of class occdat}

  \item{minimal}{Default is FALSE}
}
\value{
  An object of class npnsp (for npn spatial)
}
\description{
  Coerce elements of output from a call to occ to a single
  data.frame
}

