% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnpn-package.R
\docType{data}
\name{taxonlist}
\alias{taxonlist}
\title{Lookup-table for IDs of species and common names}
\format{A data.frame with 897 rows and 6 columns
\describe{
 \item{species_id}{species identifiers}
 \item{common_name}{common (vernacular) name}
 \item{genus}{genus name}
 \item{epithet}{epithet name}
 \item{itis_tsn}{ITIS taxonomic serial number (tsn)}
 \item{genus_epithet}{genus name + epithet name}
}}
\description{
Lookup-table for IDs of species and common names
}
\keyword{data}
