% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getindspatstations}
\alias{getindspatstations}
\title{Get all observations for a particular species or set of species.}
\usage{
getindspatstations(speciesid, stationid, year = NULL, ...)
}
\arguments{
\item{speciesid}{Required. Species id numbers, from 1 to infinity, potentially,
use e.g., c(52, 53, etc.) if more than one species desired (numeric)}

\item{stationid}{Required. Use e.g., c(4881, 4882, etc.) if more than one species desired
(numeric)}

\item{year}{Year (numeric).}

\item{...}{Optional additional curl options (debugging tools mostly)}
}
\value{
Observations for each species by date.
}
\description{
Get all observations for a particular species or set of species.
}
\examples{
\dontrun{
getindspatstations(speciesid = 35, stationid = c(60, 259), year = 2009)
getindspatstations(35, c(60, 259), 2009, 'xml')
}
}

