% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getobsspbyday}
\alias{getobsspbyday}
\title{Get observations by day for a particular species or set of species.}
\usage{
getobsspbyday(speciesid = NULL, startdate = NULL, enddate = NULL, ...)
}
\arguments{
\item{speciesid}{Required. Species id numbers, from 1 to infinity, potentially,
use e.g., c(52, 53, etc.) if more than one species desired (numeric)}

\item{startdate}{start date of data period desired, see format in examples (character)}

\item{enddate}{end date of data period desired, see format in examples (character)}

\item{callopts}{Optional additional curl options (debugging tools mostly)}
}
\value{
Number of observations by day, in object of class npn.
}
\description{
Get observations by day for a particular species or set of species.
}
\examples{
\dontrun{
out <- getobsspbyday(speciesid=357, startdate='2010-04-01', enddate='2012-01-05')
head(out[[1]])

# Lookup names
temp <- lookup_names(name='bird', type='common')
comnames <- temp[temp$species_id \%in\% c(357, 359, 1108), 'common_name']

out <- getobsspbyday(speciesid=c(357, 359, 1108), startdate='2010-04-01', enddate='2013-09-31')
names(out) <- comnames
df <- ldply(out)
df$date <- as.Date(df$date)

library('ggplot2')
ggplot(df, aes(date, count)) +
 geom_line() +
 theme_grey(base_size=20) +
 facet_grid(.id ~.)
}
}

