% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getobsspindloc}
\alias{getobsspindloc}
\title{Return all of the data, positive or negative, for an individual or a group
   of individuals, which are of a common species, at any number of locations.}
\usage{
getobsspindloc(stationid, speciesid, year = 2010, individualid = NULL, ...)
}
\arguments{
\item{stationid}{Required. Station ID; Use e.g., c(4881, 4882, etc.) if more than
one station desired (numeric).}

\item{speciesid}{Required. Species ID number (numeric).}

\item{year}{Required. Year (numeric). Default: 2010}

\item{individualid}{An individual's ID. (numeric)}

\item{...}{Optional additional curl options (debugging tools mostly)}
}
\value{
Data frame
}
\description{
Return all of the data, positive or negative, for an individual or a group
   of individuals, which are of a common species, at any number of locations.
}
\examples{
\donttest{
getobsspindloc(stationid = c(4881, 4882), speciesid = 3, year = 2009)
getobsspindloc(stationid = c(4881, 4882), speciesid = 3)
getobsspindloc(stationid = 4881, speciesid = 67)
}
}

