% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getstationsbystate}
\alias{getstationsbystate}
\title{Get number of stations by state.}
\usage{
getstationsbystate(...)
}
\arguments{
\item{...}{Optional additional curl options (debugging tools mostly)}
}
\value{
Number of stations by state as a data.frame.
}
\description{
Get number of stations by state.
}
\examples{
\donttest{
head( getstationsbystate() )
}
}

