% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getstationswithspp}
\alias{getstationswithspp}
\title{Get a list of all stations which have an individual whom is a member of a
   set of species.}
\usage{
getstationswithspp(speciesid, ...)
}
\arguments{
\item{speciesid}{Required. Species id numbers, from 1 to infinity, potentially,
use e.g., c(52, 53, etc.) if more than one species desired (numeric)}

\item{...}{Optional additional curl options (debugging tools mostly)}
}
\value{
Stations' latitude and longitude, names, and ids.
}
\description{
Get a list of all stations which have an individual whom is a member of a
   set of species.
}
\examples{
\dontrun{
getstationswithspp(speciesid = c(52,53,54))
getstationswithspp(speciesid = 53)
}
}

