% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npn_geoserver.R
\name{npn_download_geospatial}
\alias{npn_download_geospatial}
\title{Download Geospatial Data}
\usage{
npn_download_geospatial(coverage_id, date, format = "geotiff",
  output_path = NULL)
}
\value{
Data frame containing all layer details as specified in function description.
}
\description{
Function for directly downloading any arbitrary Geospatial layer data from the NPN Geospatial web services.
}
\details{
Information about the layers can also be viewed at the getCapbilities page directly:
 https://geoserver.usanpn.org/geoserver/wms?request=GetCapabilities
}
\examples{
\dontrun{
ras<-npn_download_geospatial("si-x:30yr_avg_six_bloom","255")
}
}
