% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npn_geoserver.R
\name{npn_get_custom_agdd_raster}
\alias{npn_get_custom_agdd_raster}
\title{Get Custom AGDD Raster Map}
\usage{
npn_get_custom_agdd_raster(
  method,
  climate_data_source,
  temp_unit,
  start_date,
  end_date,
  base_temp,
  upper_threshold = NULL
)
}
\arguments{
\item{method}{Takes \code{"simple"} or \code{"double-sine"} as input. This is the AGDD
calculation method to use for each data point. Simple refers to simple
averaging.}

\item{climate_data_source}{Specified the climate data set to use. Takes
either \code{"PRISM"} or \code{"NCEP"} as input.}

\item{temp_unit}{The unit of temperature to use in the calculation. Takes
either \code{"Fahrenheit"} or \code{"Celsius"} as input.}

\item{start_date}{Date at which to begin the AGDD calculations.}

\item{end_date}{Date at which to end the AGDD calculations.}

\item{base_temp}{This is the lowest temperature for each day  for it to be
considered in the calculation.}

\item{upper_threshold}{This parameter is only applicable for the double-sine
method. This sets the highest temperature to be considered in any given
day's AGDD calculation.}
}
\value{
A \link[terra:SpatRaster-class]{terra::SpatRaster} object of each calculated AGDD numeric values
based on specified time period/method/base temp/data source.
}
\description{
This function takes a series of variables used in calculating AGDD and
returns a raster of the continental USA with each pixel representing the
calculated AGDD value based on start and end date. This function leverages
the USA-NPN geo web services.
}
\examples{
\dontrun{
res <- npn_get_custom_agdd_raster(
  method = "simple",
  climate_data_source = "NCEP",
  temp_unit = "Fahrenheit",
  start_date = "2020-01-01",
  end_date = "2020-01-15",
  base_temp = 32
)
}
}
