% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npn_data_download.R
\name{npn_get_download_url}
\alias{npn_get_download_url}
\title{Generate Download URL}
\usage{
npn_get_download_url(endpoint)
}
\arguments{
\item{endpoint}{The service point, e.g. "observations/getObservations.json?"}
}
\value{
The URL, as a string.
}
\description{
Utility function to create the service point URL. Base URL comes from zzz.R,
endpoint is specified in the code. This function will manually put those
query parameters into the proper GET syntax.
}
\examples{
\dontrun{
url <- npn_get_download_url("/observations/getMagnitudeData.ndjson")
}
}
\keyword{internal}
