% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npn_phenophases.R
\name{npn_get_phenophases_for_taxon}
\alias{npn_get_phenophases_for_taxon}
\title{Get Phenophases for Taxon}
\usage{
npn_get_phenophases_for_taxon(
  family_ids = NULL,
  order_ids = NULL,
  class_ids = NULL,
  genus_ids = NULL,
  date = NULL,
  return_all = 0,
  ...
)
}
\arguments{
\item{family_ids}{Integer vector of taxonomic family ids to search for.}

\item{order_ids}{Integer vector of taxonomic order ids to search for.}

\item{class_ids}{Integer vector of taxonomic class ids to search for}

\item{genus_ids}{Integer vector of taxonomic genus ids to search for}

\item{date}{Specify the date of interest. For this function to return
anything, either this value must be set or \code{return_all} must be \code{1}.}

\item{return_all}{Takes either \code{0} or \code{1} as input and defaults to \code{0}. For
this function to return anything, either this value must be set to \code{1} or
\code{date} must be set.}

\item{...}{Currently unused.}
}
\value{
A data frame listing phenophases in the NPN database for the
specified taxon and date.
}
\description{
This function gets a list of phenophases that are applicable for a provided
taxonomic grouping, e.g. family, order. Note that since a higher taxononmic
order will aggregate individual species not every phenophase returned through
this function will be applicable for every species belonging to that
taxonomic group.
}
\details{
It's also important to note that phenophase definitions can change for
individual species over time, so there's a need to specify either a date of
interest, or to explicitly state that the function should return all
phenophases that were ever applicable for any species belonging to the
specified taxonomic group.

When called, this function requires of these three parameters, exactly one of
\code{family_ids}, \code{order_ids} or \code{class_ids} to be set.
}
