% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npn_partner_groups.R
\name{npn_groups}
\alias{npn_groups}
\title{Get Partner Groups}
\usage{
npn_groups(use_hierarchy = FALSE, ...)
}
\arguments{
\item{use_hierarchy}{Boolean indicating whether or not the list of networks should be represented in a hierarchy. Defaults to FALSE}

\item{...}{Curl options passed on to \code{\link[httr]{GET}}}
}
\value{
List of partner groups, including ID and name
}
\description{
Returns a list of all groups participating in the NPN's data collection program. These details can be used to further filter
other service endpoints' results.
}
