% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npn_obscomm.R
\name{npn_obscomm}
\alias{npn_obscomm}
\title{Get observation comments for a particular species or set of species.}
\usage{
npn_obscomm(observationid, ...)
}
\arguments{
\item{observationid}{Required. Observation id, or as vector if >1, as e.g., c(2, 1938) (numeric)}

\item{...}{Curl options passed on to \code{\link[httr]{GET}}}
}
\value{
Comments.
}
\description{
Get observation comments for a particular species or set of species.
}
\examples{
\dontrun{
npn_obscomm(c(1,4,5,7,89))
}
}

