% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npn_phenophases.R
\name{npn_pheno_classes}
\alias{npn_pheno_classes}
\title{Get Pheno Classes}
\usage{
npn_pheno_classes(...)
}
\arguments{
\item{...}{Currently unused.}
}
\value{
A tibble listing the pheno classes in the NPN database.
}
\description{
Gets information about all pheno classes, which are a higher-level order of
phenophases.
}
\examples{
\dontrun{
pc <- npn_pheno_classes()
}
}
