% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npn_phenophases.R
\name{npn_phenophase_details}
\alias{npn_phenophase_details}
\title{Get Phenophase Details}
\usage{
npn_phenophase_details(ids = list(), ...)
}
\arguments{
\item{ids}{Takes a list of phenophase ids for which to retrieve additional details.}

\item{...}{Curl options passed on to \code{\link[httr]{GET}}}
}
\description{
Retrieves additional details for select phenophases, including full list of applicable phenophase definition IDs and phenophase
revision notes over time
}
