% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npn_get_species.R
\name{npn_species_types}
\alias{npn_species_types}
\title{Get Species Types}
\usage{
npn_species_types(kingdom = c("Plantae", "Animalia"), ...)
}
\arguments{
\item{kingdom}{Filters results by taxonomic kingdom. Valid values include
\code{'Animalia'}, \code{'Plantae'}, or \code{NULL} (which returns results for both).
Defaults to \code{'Plantae'}.}

\item{...}{Currently unused.}
}
\value{
A data frame with a list of the functional types used in the NPN
database, filtered by the specified kingdom.
}
\description{
Return all plant or animal functional types used in the NPN database.
}
\examples{
\dontrun{
npn_species_types("Plantae")
}
}
