% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npn_get_species.R
\name{npn_species_types}
\alias{npn_species_types}
\title{Get Species Types}
\usage{
npn_species_types(kingdom = "Plantae", ...)
}
\arguments{
\item{kingdom}{The kingdom for which to return functional types; either 'Animalia' or 'Plantae'. Defaults to Plantae.}
}
\description{
Return all plant or animal functional types used in the NPN database.
}
