% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npn_stationsbystate.R
\name{npn_stationsbystate}
\alias{npn_stationsbystate}
\title{Get number of stations by state.}
\usage{
npn_stationsbystate(...)
}
\arguments{
\item{...}{Curl options passed on to \code{\link[httr]{GET}}}
}
\value{
Number of stations by state as a data.frame.
}
\description{
Get number of stations by state.
}
\examples{
\dontrun{
head( npn_stationsbystate() )
}
}
