% Generated by roxygen2 (4.0.2): do not edit by hand
\name{npn_todf}
\alias{npn_todf}
\title{Coerce elements of an object of class npn to a single data.frame}
\usage{
npn_todf(input, minimal = FALSE)
}
\arguments{
\item{minimal}{Default is FALSE}

\item{x}{An object of class npn}
}
\value{
An object of class npnsp (for npn spatial), containing just a data.frame
   with the following fields if minimal=TRUE
\enumerate{
  \item sciname
  \item latitude
  \item longitude
}

and the following fields if minimal=FALSE
\enumerate{
  \item sciname
  \item latitude
  \item longitude
  \item asdfas
  \item asdfdh
  \item asdfawer
  \item asdfssawe
  \item asdfad
  \item asdfwe
  \item adsfcwer
}
}
\description{
Coerce elements of an object of class npn to a single data.frame
}

