% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defunct.R
\name{rnpn-defunct}
\alias{rnpn-defunct}
\title{Defunct functions in rnpn}
\description{
\itemize{
 \item \code{\link{npn_obsspbyday}}: Removed.
 \item \code{\link{npn_allobssp}}: Removed.
 \item \code{\link{npn_indspatstations}}: Removed.
 \item \code{\link{npn_indsatstations}}: Removed.
 \item \code{\link{npn_stationsbystate}}: Removed.
}
}
