% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnpn-package.R
\docType{package}
\name{rnpn-package}
\alias{rnpn-package}
\alias{rnpn}
\title{Interface to the National Phenology Network API}
\description{
This package allows for easy access to the National Phenology
Network's Data API. To learn more, take a look at the vignettes.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/usa-npn/rnpn}
  \item \url{http://usa-npn.github.io/rnpn/}
  \item Report bugs at \url{https://github.com/usa-npn/rnpn/issues}
}

}
\author{
\strong{Maintainer}: Jeff Switzer \email{jeff@usanpn.org}

Authors:
\itemize{
  \item Jeff Switzer \email{jeff@usanpn.org}
  \item Scott Chamberlain
  \item Lee Marsh
  \item Kevin Wong
  \item Eric R Scott (\href{https://orcid.org/0000-0002-7430-7879}{ORCID})
}

Other contributors:
\itemize{
  \item David LeBauer [contributor]
}

}
\keyword{package}
