% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnpn-package.R
\docType{package}
\name{rnpn-package}
\alias{rnpn-package}
\alias{rnpn}
\title{Interface to the National Phenology Network API}
\description{
This package allows for easy access to the National Phenology Network's Data API. To learn more, take a look at the vignettes.
}
\keyword{package}
